refactor: getInvolved card to components directory
